requirements.txt
numpy
pandas
matplotlib
pyyaml
brokenaxes
pyserial
sympy
scipy
scikit-learn
plotly
dash